// src/app/languages/interface/more/shop.interface.ts
export interface IShop {
    dealsOftheDay: string; 
    buyXGetY: string;
    freeGifts: string;
    categories: string;
    topRated: string;
    mostSearched: string;
    recommended: string;
    recentlyViewed: string;
    couponProducts: string;
    justLaunched: string;
    yourBrowsingHistory: string;
}